requirements.txt
jinja2
click
pyyaml
